refactor: split mcp-server/utils.ts into focused modules (#112)
- shared.ts: ROOT, readAiOsFile, path helpers, lock utilities, I/O helpers
- memory.ts: repo-memory store and hygiene (getRepoMemory, rememberRepoFact, pruneMemory, syncHostedMemory, runMemoryMaintenance)
- session.ts: active plan, checkpoints, failure ledger, watchdog
- search.ts: searchFiles, buildFileTree
- project-introspection.ts: getPrismaSchema, getTrpcProcedures, getApiRoutes, getEnvVars, getPackageInfo, getFileSummary, getImpactOfChange, getDependencyChain
- freshness-bridge.ts: getContextFreshness (wraps detectors/freshness)
- recommendations-bridge.ts: getRecommendations, suggestImprovements
- utils.ts: thin barrel (133 lines) with 4 real impls + re-exports for backward compat
- tests/mcp-server-modules.test.ts: export-contract + runtime smoke tests

Build: zero TS errors | Tests: 215/215 passed
